Move loginValidators and registerValidators to utils.js. Export and import validators where needed.
