+++ routes/users.js
@@ -7,7 +7,11 @@ const {
 const db = require("../db/models");
 const {
   csrfProtection,
-  asyncHandler
+  asyncHandler,
+  getSingleWheelhouse,
+  getAllWheelhouses,
+  registerValidators,
+  loginValidators,
 } = require("./utils.js");
 const {
   check,
@@ -20,91 +24,6 @@ const router = express.Router();
 // router.get('/', function(req, res, next) {
 //   res.send('respond with a resource');
 // });
-
-const registerValidators = [
-  check("firstName")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please enter your first name")
-  .isLength({
-    max: 50
-  })
-  .withMessage("First name cannot be more than 50 characters long."),
-  check("lastName")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please enter your last name")
-  .isLength({
-    max: 50
-  })
-  .withMessage("Last name cannot be more than 50 characters long."),
-  check("email")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please enter your email address.")
-  .isEmail({
-    checkFalsy: true
-  })
-  .withMessage("Please enter a valid email address.")
-  .isLength({
-    max: 50
-  })
-  .withMessage("Email address cannot be more than 50 characters long.")
-  .custom((email) => {
-    return db.User.findOne({
-      where: {
-        email
-      }
-    }).then((user) => {
-      if (user) return Promise.reject("Email address already in use.");
-    });
-  }),
-  check("username")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please provide a username.")
-  .isLength({
-    max: 50
-  })
-  .withMessage("Username cannot be more than 50 characters long.")
-  .custom((username) => {
-    return db.User.findOne({
-      where: {
-        username
-      }
-    }).then((user) => {
-      if (user)
-        return Promise.reject(
-          "Username unavailable, please try something different."
-        );
-    });
-  }),
-  check("password")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please enter a password.")
-  .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])/, "g")
-  .withMessage(
-    "Password must contain uppercase letter, lowercase letter, a number, and a special character."
-  ),
-  check("confirmPassword")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please confirm your password.")
-  .custom((confirmPassword, {
-    req
-  }) => {
-    if (confirmPassword !== req.body.password)
-      throw new Error("Please make sure your passwords match");
-    return true;
-  }),
-];
 
 router.get(
   "/register",
@@ -189,19 +108,6 @@ router.get(
     });
   })
 );
-
-const loginValidators = [
-  check("username")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please provide your username."),
-  check("password")
-  .exists({
-    checkFalsy: true
-  })
-  .withMessage("Please provide your password."),
-];
 
 
 router.post(
